# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Eventify API",
    "description": "Real-time event tracking and analytics platform with WebSocket support, gRPC streaming, and comprehensive dashboard metrics.",
    "version": "1.0.0",
    "contact": {
      "name": "Eventify Team",
      "email": "support@eventify.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api-key": {
      "get": {
        "tags": ["API Keys"],
        "summary": "Get user API keys",
        "description": "Retrieve all API keys for the authenticated user",
        "operationId": "getUserApiKeys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "API keys retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ApiKey"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["API Keys"],
        "summary": "Create API key",
        "description": "Create a new API key for the authenticated user",
        "operationId": "createApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKey"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api-key/validate": {
      "get": {
        "tags": ["API Keys"],
        "summary": "Validate API key",
        "description": "Validate an API key",
        "operationId": "validateApiKey",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key validation result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKeyValidationResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api-key/{key}": {
      "delete": {
        "tags": ["API Keys"],
        "summary": "Delete API key",
        "description": "Delete an API key",
        "operationId": "deleteApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "API key to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "API key deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["API Keys"],
        "summary": "Update API key",
        "description": "Update API key activation status",
        "operationId": "updateApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "description": "API key to update",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateApiKeyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "API key updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiKey"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "API key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Format: Bearer <token>"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for event ingestion"
      }
    },
    "schemas": {
...
      "CreateApiKeyDto": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Name for the API key",
            "example": "Production API Key"
          }
        }
      },
      "ValidateApiKeyDto": {
        "type": "object",
        "required": ["apiKey"],
        "properties": {
          "apiKey": {
            "type": "string",
            "description": "API key to validate",
            "example": "ak_1234567890abcdef"
          }
        }
      },
      "UpdateApiKeyDto": {
        "type": "object",
        "required": ["isActive"],
        "properties": {
          "isActive": {
            "type": "boolean",
            "description": "Whether the API key should be active",
            "example": true
          }
        }
      },
...
      "ApiKey": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "API key ID",
            "example": "64a1b2c3d4e5f6789012"
          },
          "name": {
            "type": "string",
            "description": "API key name",
            "example": "Production API Key"
          },
          "key": {
            "type": "string",
            "description": "API key value",
            "example": "ak_1234567890abcdef"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the API key is active",
            "example": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "API key creation timestamp",
            "example": "2024-08-08T10:30:00Z"
          },
          "lastUsedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last usage timestamp",
            "example": "2024-08-08T12:30:00Z"
          }
        }
      },
      "ApiKeyValidationResponse": {
        "type": "object",
        "properties": {
          "valid": {
            "type": "boolean",
            "description": "Whether the API key is valid",
            "example": true
          },
          "ownerId": {
            "type": "string",
            "description": "Owner user ID",
            "example": "64a1b2c3d4e5f6789012"
          },
          "keyName": {
            "type": "string",
            "description": "API key name",
            "example": "Production API Key"
          }
        }
      },
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation completed successfully"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "example": "Invalid input data"
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          }
        }
      }
    }
  },
...
}